requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.0
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.8
decorator==5.1.1
fonttools==4.43.1
freelunch==0.0.15
GPy==1.13.1
idna==3.4
jax==0.4.20
jaxlib==0.4.20
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.8.0
ml-dtypes==0.3.1
mpmath==1.3.0
numpy==1.26.1
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.1
paramz==0.9.6
Pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2024.1
requests==2.31.0
scikit-learn==1.3.1
scipy==1.11.3
six==1.16.0
sympy==1.12
threadpoolctl==3.2.0
tzdata==2024.1
urllib3==2.0.6
